synrm_v2: MTPA LUT -> float32 amps (full +-450 A range)
int16 (even centiamps) tops out at +-327 A, too small for the +-450 A
controller (the final motor may pull ~200 A; id* can exceed -327 A).
Store foc_mtpa_lut[] as float (AMPS) instead: no scale, no overflow,
covers the full hardware range and any future table.

- datatypes.h: foc_mtpa_lut[] int16 -> float (amps).
- confgenerator.c: serialize/deserialize via buffer_append/get_float32_auto;
  set_defaults local array -> float. confgenerator.h signature
  729934208 -> 716829362 (matches vesc_tool XML).
- mcconf_default.h: default table in amps; foc_mtpa_lut_imax 70 -> 81 A
  (the 9kW map's full current axis).
- mcpwm_foc.c: synrm_mtpa_id() returns the amp value directly (no *scale).

foc_mtpa_lut_imax spans the full characterized current axis, capped at the
450 A controller limit. Serialized mcconf ~655 B, under 1024 B. Gotcha G16.
Builds clean on staryag2 + staryag3.

+++ datatypes.h
@@ -489,8 +489,8 @@ typedef struct {
 	float foc_synrm_erpm_01;
 	float foc_synrm_erpm_12;
 	float foc_synrm_blend;
-	int16_t foc_mtpa_lut[MTPA_LUT_SIZE]; // SynRM MTPA id*(|I|) in milliamps, uniform |I| axis 0..foc_mtpa_lut_imax
-	float foc_mtpa_lut_imax;             // current magnitude (A) at the last LUT sample
+	float foc_mtpa_lut[MTPA_LUT_SIZE];   // SynRM MTPA id*(|I|) in AMPS, uniform |I| axis 0..foc_mtpa_lut_imax
+	float foc_mtpa_lut_imax;             // current magnitude (A) at the last LUT sample (up to the 450 A controller limit)
 	float foc_motor_r;
 	float foc_motor_flux_linkage;
 	float foc_observer_gain;
